+++ packages/blog/src/hooks/useSelectedTag/index.ts
@@ -1,0 +1,1 @@
+export { default } from './useSelectedTag';
