+++ packages/blog/src/hooks/useSelectedTag/useSelectedTag.ts
@@ -1,0 +1,21 @@
+import { useState } from 'react';
+
+const useSelectedTag = () => {
+  const [selectedTag, setSelectedTag] = useState<string>('');
+
+  const handleSelected = (tag: string) => {
+    if (!selectedTag || selectedTag !== tag) {
+      setSelectedTag(tag);
+      return;
+    }
+
+    setSelectedTag('');
+  };
+
+  return {
+    selectedTag,
+    handleSelected,
+  };
+};
+
+export default useSelectedTag;
